+++ src/components/navigation/mobile-nav.tsx
@@ -58,15 +58,15 @@ const navigation: NavItem[] = [
     ),
   },
   {
-    name: 'Messages',
-    href: '/messages',
+    name: 'SMS Templates',
+    href: '/sms-templates',
     icon: (props) => (
       <svg fill="none" viewBox="0 0 24 24" stroke="currentColor" {...props}>
         <path
           strokeLinecap="round"
           strokeLinejoin="round"
           strokeWidth={2}
-          d="M8 10h.01M12 10h.01M16 10h.01M9 16H5a2 2 0 01-2-2V6a2 2 0 012-2h14a2 2 0 012 2v8a2 2 0 01-2 2h-5l-5 5v-5z"
+          d="M4 6h16M4 10h16M4 14h16M4 18h16"
         />
       </svg>
     ),