+++ src/components/navigation/sidebar.tsx
@@ -79,51 +79,6 @@ const navigation: NavItem[] = [
     mobileOnly: false
   },
   {
-    name: 'Customers',
-    href: '/customers',
-    icon: (props) => (
-      <svg fill="none" viewBox="0 0 24 24" stroke="currentColor" {...props}>
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth={2}
-          d="M12 4.354a4 4 0 110 5.292M15 21H3v-1a6 6 0 0112 0v1zm0 0h6v-1a6 6 0 00-9-5.197M13 7a4 4 0 11-8 0 4 4 0 018 0z"
-        />
-      </svg>
-    ),
-    mobileOnly: false
-  },
-  {
-    name: 'Categories',
-    href: '/categories',
-    icon: (props) => (
-      <svg fill="none" viewBox="0 0 24 24" stroke="currentColor" {...props}>
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth={2}
-          d="M7 7h.01M7 3h5c.512 0 1.024.195 1.414.586l7 7a2 2 0 010 2.828l-7 7a2 2 0 01-2.828 0l-7-7A1.994 1.994 0 013 12V7a4 4 0 014-4z"
-        />
-      </svg>
-    ),
-    mobileOnly: false
-  },
-  {
-    name: 'Messages',
-    href: '/messages',
-    icon: (props) => (
-      <svg fill="none" viewBox="0 0 24 24" stroke="currentColor" {...props}>
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth={2}
-          d="M8 10h.01M12 10h.01M16 10h.01M9 16H5a2 2 0 01-2-2V6a2 2 0 012-2h14a2 2 0 012 2v8a2 2 0 01-2 2h-5l-5 5v-5z"
-        />
-      </svg>
-    ),
-    mobileOnly: false
-  },
-  {
     name: 'SMS Templates',
     href: '/sms-templates',
     icon: (props) => (